adjust css
