+++ src/routes/result/result.css
@@ -1,3 +1,9 @@
+#result {
+  display: grid;
+  grid-template-columns: repeat(auto-fit, minmax(320px, 1fr));
+  grid-gap: 1em;
+}
+
 #result-table {
   width: 100%;
   border-spacing: 0;
